Added restarting story to after scenario annotated method
Added support for restarting story in before/after annotated steps

JBEHAVE-1053: Fixed restarting story to allow restarts when the story
ends with a pending step and inside of an examles table

Keep track of recursive calls to run(), so the correct delegete is
executed to log a restarted story while multi-threading.

Cleaned up commented out code. Report restarted stories at end of the
story run, so when multi-threading it will report properly

Added try/catch to ensure the storyRestartDepth is decremented even if failures occur